feat(54): paging and sorting implementation

+++ src/main/java/com/compstore/dto/PagingAndSortingMetadataDTO.java
@@ -11,5 +11,5 @@ public class PagingAndSortingMetadataDTO {
 
     private Integer pagesCount;
 
-    private Integer elementsCount;
+    private Long elementsCount;
 }
